Extract datamodel to its own class
